requirements.txt
scikit-learn==1.3.1
fastapi
numpy
uvicorn
pydantic
python-whois
dnspython
requests
